+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.74.1
 
+-   backups: Add integrityCheck to LocatorInfo